refactor: Update Notion types and improve post filtering logic
- Refactored Notion-related types by moving them to a new types directory for better organization.
- Introduced a new NOTION_PROPERTY constant to standardize property names used across the application.
- Enhanced the post filtering logic in the Notion service to support series filtering and improved readability.
- Added a new "series" option to the PostCategorySelector component.
- Updated the PostCard component to handle series clicks and display series tags correctly.
- Adjusted imports to reflect the new type structure.
